fix: 14个功能模块全量修复优化 - 操作链路完整性保障
基本模块: Shader rgba8归一化修复(7文件)、Alpha通道污染修复(6文件)、
白平衡b_mult公式修正、色温显示公式修正(ln替代pow)

曲线模块: toneCurvePoints计数修复、Alpha通道污染修复、
ComputeTangents死代码清除、整数溢出防护

色彩模块: channel_mixer/color_wheel输出clamp、空指针检查、
channels校验

HSL模块: 浮点比较epsilon化、除零防护、OkLab输出clamp、
M_PI可移植性修复

几何模块: 透视变换/缩放除零防护、M_PI精度修复、
Shader #version和rgba32f、CropOverlay/GeometryPanel除零守卫

效果模块: film_grain仅RGB通道、halation高光mask上界裁剪、
Shader #version和rgba32f归一化

RAW模块: highlight_reconstruction空指针/参数验证、
rcd_demosaic/highlight_recon Shader #version+除零保护

历史模块: HistoryPanel collectAsStateWithLifecycle、
compressTransactions游标映射修复、EditHistoryRepository
JSON解析修复(不再丢失Pipeline参数)、ordinal越界检查

色彩科学: NaN/Inf防护(7处)、planckian_bulk色调一致性

LMT: intensity参数传递、lut3d.comp rgba16f+三线性插值

检查器: HistogramView越界阈值、CompareView空图占位

镜头校正: 迭代逆向畸变模型、Shader越界输出黑、
midpoint保护、LensCorrectionPanel remember修复

蒙版: MaskPanel collectAsStateWithLifecycle、BrushOverlay
onDragCancel保存笔触

预设: PresetService lutIntensity默认1f、insert检查

接口一致性: sigmoidShoulder/sigmoidPivot参数偏移修复、
PipelineParams结构体添加字段、native-lib.cpp参数读取对齐、
pipeline_service.cpp sigmoid_contrast_bulk传参修复、
FilmGrainOp channels参数传递

+++ AlcedoAndroid/app/src/main/cpp/core/edit/operators/crop_rotate_op.cpp
@@ -86,7 +86,7 @@ void CropRotateOperator::apply_rotate(float* dst, const float* src, int src_widt
         }
     } else {
         // Arbitrary rotation using bilinear interpolation
-        float rad = static_cast<float>(angle) * M_PI / 180.0f;
+        float rad = static_cast<float>(angle) * static_cast<float>(M_PI) / 180.0f;
         float cos_a = std::cos(rad);
         float sin_a = std::sin(rad);
